requirements.txt
Django
djangorestframework
Pillow
psycopg2


